+++ test/lightning/extensions/external_metrics_test.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Extensions.ExternalMetricsTest do
+  use Lightning.DataCase, async: true
+
+  alias Lightning.Extensions.ExternalMetrics
+
+  test "returns an empty list " do
+    assert ExternalMetrics.plugins() == []
+  end
+end