+++ test/lightning_web/live/credential_live_test.exs
@@ -303,7 +303,7 @@ defmodule LightningWeb.CredentialLiveTest do
 
       view
       |> element("#delete_credential_#{credential.id}_modal_confirm_button")
-      |> render_click() =~ "Credential deleted successfully!"
+      |> render_click() =~ "Credential deleted"
 
       {:ok, _view, html} =
         live(conn, ~p"/projects/#{project}/settings#credentials",
@@ -986,69 +986,6 @@ defmodule LightningWeb.CredentialLiveTest do
       {_path, flash} = assert_redirect(index_live)
       assert flash == %{"info" => "Credential updated successfully"}
     end
-
-    # test "blocks credential transfer to invalid owner; allows to valid owner", %{
-    #   conn: conn,
-    #   user: first_owner,
-    #   credential: credential_1
-    # } do
-    #   user_2 = insert(:user)
-    #   user_3 = insert(:user)
-
-    #   project =
-    #     insert(:project,
-    #       name: "myproject",
-    #       project_users: [%{user_id: first_owner.id}, %{user_id: user_2.id}]
-    #     )
-
-    #   credential =
-    #     insert(:credential,
-    #       user: first_owner,
-    #       name: "the one for giving away",
-    #       project_credentials: [
-    #         %{project: project, credential: nil}
-    #       ]
-    #     )
-
-    #   {:ok, index_live, html} = live(conn, ~p"/credentials")
-
-    #   # both credentials appear in the list
-    #   assert html =~ credential_1.name
-    #   assert html =~ credential.name
-
-    #   assert html =~ first_owner.id
-    #   assert html =~ user_2.id
-    #   assert html =~ user_3.id
-
-    #   assert index_live
-    #          |> form("#credential-form-#{credential.id}",
-    #            credential: Map.put(@update_attrs, :user_id, user_3.id)
-    #          )
-    #          |> render_change() =~ "Invalid owner"
-
-    #   #  Can't transfer to user who doesn't have access to right projects
-    #   assert index_live |> submit_disabled()
-
-    #   {:ok, _index_live, html} =
-    #     index_live
-    #     |> form("#credential-form-#{credential.id}",
-    #       credential: %{
-    #         body: "{\"a\":\"new_secret\"}",
-    #         user_id: user_2.id
-    #       }
-    #     )
-    #     |> render_submit()
-    #     |> follow_redirect(
-    #       conn,
-    #       ~p"/credentials"
-    #     )
-
-    #   # Once the transfer is made, the credential should not show up in the list
-
-    #   assert html =~ "Credential updated successfully"
-    #   assert html =~ credential_1.name
-    #   refute html =~ "the one for giving away"
-    # end
   end
 
   describe "Authorizing an oauth credential" do
@@ -3149,6 +3086,27 @@ defmodule LightningWeb.CredentialLiveTest do
         |> follow_redirect(conn, ~p"/credentials")
 
       assert html =~ "Could not revoke transfer"
+    end
+  end
+
+  describe "credential type picker modal" do
+    test "displays credential type modal with icons", %{conn: conn} do
+      {:ok, _view, html} = live(conn, ~p"/credentials")
+
+      html_tree = Floki.parse_document!(html)
+
+      for adaptor <- ["postgresql", "dhis2", "http"] do
+        adaptor_label =
+          Floki.find(
+            html_tree,
+            "label[for='credential-schema-picker_selected_#{adaptor}']"
+          )
+
+        adaptor_icon = Floki.find(adaptor_label, "object")
+        assert length(adaptor_icon) > 0
+        img_src = adaptor_icon |> Floki.attribute("data") |> List.first()
+        assert img_src =~ "/images/adaptors/#{adaptor}-square.png"
+      end
     end
   end